RC_Channel: add ARM auxiliary switch option (188) and 18-channel support
Support ARM auxiliary switch function (188) and expand RC channels
for compact companion or WiFi controllers.

+++ libraries/RC_Channel/RC_Channels.cpp
@@ -181,13 +181,13 @@ void RC_Channels::clear_overrides(void)
     // copter and plane, RC_Channels needs to control failsafes to resolve this
 }
 
-uint16_t RC_Channels::get_override_mask(void) const
+uint32_t RC_Channels::get_override_mask(void) const
 {
-    uint16_t ret = 0;
+    uint32_t ret = 0;
     RC_Channels &_rc = rc();
     for (uint8_t i = 0; i < NUM_RC_CHANNELS; i++) {
         if (_rc.channel(i)->has_override()) {
-            ret |= (1U << i);
+            ret |= (1UL << i);
         }
     }
     return ret;
